test commit 4

+++ main.py
@@ -2,7 +2,7 @@ def main():
     print('My first git program')
     print("Hi!")
     print('And I change it every day')
-    x = int(input())
+    x = float(input())
 
 
 if __name__ == '__main__':
